refs #19702 Updated test

+++ Framework/Algorithms/src/MuonAsymmetryHelper.cpp
@@ -92,7 +92,7 @@ double estimateNormalisationConst(const HistogramData::Histogram &histogram,
   auto iyN = std::next(yData.rawData().begin(), iN);
   double summation = std::accumulate(iy0, iyN, 0.0);
   double denominator = 0.0;
-  for (int k = i0; k < iN; k++) {
+  for (size_t k = i0; k < iN; k++) {
 	  denominator += exp(-xData[k] / MUON_LIFETIME_MICROSECONDS);
   }
   return summation  / (denominator *= numGoodFrames);
